add style file for ErrorPage
