+++ app/src/App.tsx
@@ -17,7 +17,7 @@ import { DialogProvider } from "./hooks/useDialog";
 import { ConfigProvider } from "./hooks/config";
 
 import { UserInfoPage } from "./views/UserInfo/UserInfoPage";
-import { ErrorPage } from "./components/Error/ErrorPage";
+import { ErrorPage } from "./components/ErrorPage/ErrorPage";
 import { ContactsPage } from "./views/StaticPages/Contacts/Contacts";
 import { ExamplesPage } from "./views/StaticPages/Examples/Examples";
 import { DashboardPage } from "./views/Dashboard/Dashboard";
@@ -60,10 +60,8 @@ export default function App(): JSX.Element {
       <Route path="/settings/company-info" element={<CompanyInfoPage />} />
       <Route path="/contacts" element={<ContactsPage />} />
       <Route path="/examples" element={<ExamplesPage />} />
-
       <Route path="/privacy-policy" element={<PrivacyPolicyPage />} />
       <Route path="/terms-of-service" element={<TermsOfServicePage />} />
-
       <Route path="*" element={<ErrorPage code={"404"} />} />
       <Route path="/change-password" element={<ChangePasswordPage />} />
       <Route path="users/create" element={<InviteForm />} />
